Update mysh.c
This is a version includes working cd, pwd, and exit as built-in commands. I extensively tested them for about 30 minutes before I deemed it ok for this point. Having cd or pwd not the start of the command is properly catched by the error statements, but the error printed out by perror() is a little confusing. Test this yourself to see it properly. The redirection/piping for pwd in NYI (not yet implemented).